Refactoring #2
added io_requests.h
added io_requests.cpp

added output_ui.h
added output_ui.cpp
