+++ io_requests.h
@@ -1,0 +1,15 @@
+//============================================================================
+// Name		   : interrupts.h
+// Author	   :
+// Version	   :
+// Description : Function definitions for the IO
+//============================================================================
+
+#ifndef IO_REQUESTS_H_
+#define IO_REQUESTS_H_
+
+#include "state_queues.h"
+
+void check_io_interrupt();
+
+#endif /* IO_REQUESTS_H_ */
