+++ output_ui.h
@@ -1,0 +1,16 @@
+//============================================================================
+// Name		   : output_ui.h
+// Author	   :
+// Version	   :
+// Description : Function definitions for the output
+//============================================================================
+
+#ifndef OUTPUT_UI_H_
+#define OUTPUT_UI_H_
+
+#include "state_queues.h"
+
+void debug_print_memory();
+void debug_print_new_ready();
+
+#endif /* OUTPUT_UI_H_ */
